Move ml_selection_buffers to common.
Should fix/simplify linking of plugins, etc.
