+++ src/meshlabplugins/edit_select/edit_select.pro
@@ -4,7 +4,6 @@ HEADERS       = edit_select_factory.h \
 				edit_select.h 
 
 SOURCES       = edit_select_factory.cpp \
-				../../meshlab/ml_selection_buffers.cpp \
 				edit_select.cpp
 
 TARGET        = edit_select
